Removes unnecessary bools from VarDeclInfo
As far as I know, statics and globals are the same thing in DM, grammatically synonymous. So there's no reason to maintain a separate flag.

Further, IsTopLevel was also unnecessary, since we can just automatically convert root-level var declares into being static (that was sort of the point of marking them anyways). So that boolean is gone, too.