Fixes duplicate definitions of statics and object properties not causing a compile-time error (#771)
* Removes duplicate definitions in DMStandard

This was spotted by the new dup-definition detection that statics and properties now have.

* Removes unnecessary bools from VarDeclInfo

As far as I know, statics and globals are the same thing in DM, grammatically synonymous. So there's no reason to maintain a separate flag.

Further, IsTopLevel was also unnecessary, since we can just automatically convert root-level var declares into being static (that was sort of the point of marking them anyways). So that boolean is gone, too.

* Fixes duped definitions of statics/properties not erroring

I'm not completely satisfied with this, since this still means that dupe detection is happening in two places in the codebase (statics/properties in DMObjectBuilder and all others in DMProcBuilder), but I don't think it's too much of a problem.

* Adds HasGlobalVariable, fixes dupe checks in VarDefinition

Would've forced-pushed but these are slightly non-trivial changes so

EDIT: Makes objects ignore their root globals (since those are kinda in their own namespace in a certain way)
EDIT EDIT: Reduces code duplication in ProcessVarDefinition
EDIT EDIT EDIT: Remove dumb comment