Adds a basic RAII-like defer pattern in ODS
I don't entirely trust this pattern, nor can I really commend its beauty, but we are in desperate need of a method to defer things throughout DMParser. Many functions have clunky syntax or erratic behaviour as a result of not always doing a task on all possible control paths. I've seen a couple TODOs about it as well so, here ya go.