Implements the paren-less sleep call
"sleep 40" is now a valid DM statement.

sleep, like del, has a special snowflake behaviour that allows it to be called without parentheses when acting as a lone expression-as-statement.

"sleep" however isn't precisely a keyword, at least not in OD, so this is adds a bit of snowflake flow control to interpret it correctly anyways.