Fixes several bugs with string macros, implements all pronoun/article ones (#777)
* Moves String Token parsing into a helper proc

DMParser.cs is >2300 lines long before this change, and implementing pronouns is going to make it even longer, so I figure its time to put the helper file to some use.

Marked as a separate commit so you have better diffs to look at during review :^)

* Parser now ensures that string macros have interp value if needed

A dangling \hers or \The now causes a parsing error, as it should.

* Improves quality of string macro errors/warnings

We now, like BYOND, give a descriptive error message when user tries to use \her (a macro that can't exist).

While I was here, I decided to be even sweeter and add some leniency with unnecessarily-capitalized macros. :blush:

* Refactors StringFormatType into a static class thing

Storing all of our formatting bytes as 0x00FF 0x00-whatever was both breaking parity and wasting bytes, so that's been completely changed in favour of a static class that handles encoding and decoding from formatting bytes, stored passive-aggressively into a not-as-important Unicode codeblock. Various other changes were made to prepare for a wider implementation of text macros.

This compiled and ran OK, so I'm committing here before I start implementing novel functionality.

* Adds null opcode checking in InternalResume()

I had a really strange bug recently and this helped me with debugging, so might as well commit it to master.

* Implements \the, \The, \A, \An

This gets the general pattern going for formatted strings.

At first, I thought that I was going to have to do two passes on the string to actually parse it correctly (to harvest the locations of all the interp values, so I know who owns what macro). Realized that I could just push another operand to the opcode proc and save a lot of that time, so that's what's here.

* Adds unimplemented warnings for \s and \th

Previously, they would actually cause a compiletime error were they to appear. Now the compiler treats them as it should.

* Implements suffix pronouns (\he, \himself, etc)

I need a better place to toss all these helpers.
EDIT: Accidentally left debug logging in there, whoops

* Re-implements new behaviour from #801, #720

I moved this code over into a helper proc so this new stuff has to be re-added here, awkwardly.

EDIT: Also patched for some changes caused by refs becoming semi-implemented.

* Adds some new unit tests for string interpolation

EDIT: adds StringInterpolation4.dm
EDIT EDIT: and the fifth test too

* Fixes weird bug with reversed string interp values

* Fixes whitespace not making strings proper

* Fixes \ref[] interpolations being used by suffix pronouns

* Wixoa review commit