Fixes initial() not respecting insincere var-typing
Vars that are typed can actually store ANYTHING. The type is a weak promise that BYOND uses as a guarantee in some contexts (as in statics and constants) but not in others (as in normal variable accesses).

It's still pretty sour to me that we're supporting this, but that is indeed the BYOND behaviour.