Fixes infinite spin / memory usage bug when a braced block can't find its ending brace (#966)
* Fixes infinite parser bug when brace block has no }

I noticed this, not in Yogs but in a stripped version of it where I had accidentally caused a brace block to not have its closing brace.

This resulted in the parser spinning indefinitely, creating a 200 MB/s memory leak. This fixes that :^)

* Caps the number of emissions parsers can record at 50 million

Hopefully this prevents this sort of leakage happening in the future, at least a little.