Writes native impl helpers for orange, range, view, oview
These procs are very similar, so I've set up this impl such that they all can use the same one.

Since the helper returns an enumerable, it may be possible later to avoid the allocation of a new DM /list, such as in situations like foo(var/foo in range(5)) or whatever.