Allows for fallible access of atom appearances
While implementing more of /atom.color, I ended up creating a cyclical dependency where, creating an appearance required accessing the color var, and doing a get on the color var required the appearance. This caused a stack overflow as it attempted to create an appearance a zillion times.

None of that is featured here though, just some changes to avoid the implicit instantiation of an appearance when just asking if it exists or not.