Makes OD accept oddball phrases like "if();" and "while()."
This is another pragma option, the first one to be a notice under the default config.

Apparently DM supports adding some random extra characters to the end of some keywords and block inits, some of which we already supported but not in a consistent way. This addresses all of that.