Fixes lateral proc overrides, OD now errors on missing types promised by var definitions, and more! (#934)
* Makes OD compile-time on missing types promised in vardefs

This is slightly off-parity since BYOND actually does let you do this if the variable is never provably referenced or instantiated at compile-time. However, it does fix a broken test!

Several targets now error because of this. Looks like we'll have to do another audit on them.

* Allows 'do', 'while', and 'for' to be in type paths, per parity

It took me a bit to realize why my do-re-mi test was not working :^)

Also dedents a bunch of path parsing code in favour of an early-return pattern.

* Adds better support for lateral proc overrides and global overrides

OD is now much better about warning when procs are not defined or overdefined, and now handles overrides in a completely correct way, AFAIK.

Also fixed a bug where indent-defined procs (of the form proc\n\tfoo() ) did not know they were not proc overrides.

* Removes bad global proc test

It is actually possible (albeit buggy) to override a global proc. This test doesn't really make sense and seems to be more of a BYOND bug rather than any good show of behaviour.

* Creates three new proc unit tests, promotes a broken test

OD now passes over 250 tests! Wow!

EDIT: adds a missing newline at EOF (also trying to bump the CI)

* Refactors override errors to use pragmas

Adds two new warning codes, as well :)

* Small amendment for conflict resolution

IsRoot is probably a better name anyways. I don't even like Guardians of the Galaxy.

* Demotes DanglingVarType to warning in default config

This is a parity thing - this is obviously erroneous code, but we need things like this in order to compile some codebases correctly!

* Does the lateral proc PR wixoa review

* Disables overriding of global procs, adds pragma emit for it