Implements color matrices, allows `/atom.color` to be set with one (#1062)
* Implements a new struct: ColorMatrix

This is going to end up used a lot for our implementation of DM color matrices.

* Allows for fallible access of atom appearances

While implementing more of /atom.color, I ended up creating a cyclical dependency where, creating an appearance required accessing the color var, and doing a get on the color var required the appearance. This caused a stack overflow as it attempted to create an appearance a zillion times.

None of that is featured here though, just some changes to avoid the implicit instantiation of an appearance when just asking if it exists or not.

* Adds a new shader for color matrices

* Implements setting /atom.color to a matrix, mostly

* Adds an IntegrationTest proc for /atom.color matrices

* Fixes color matrix shader, TryParseColor respects defaultAlpha

Previously, defaultAlpha would not kick in if the color was long enough. Fixed that! :^)

(hopefully that doesn't break anything :c)

* Adds color filter test to the test game

Still need to allow invokation of it via setting the filters var but, hey, look at that! :^)

* Fixes mis-transposed color matrix in colormatrix shader

Also corrects the default argument for the shader, so the client doesn't crash :^)

* Fixes compiletime-set /atom.color not being visible at runtime

* Implements the rest of the color matrix constructors

* Adds serialization of DreamValue => ColorMatrix so filter() works

This is basically what it means to implement the new color filter for filter() and the filters metaobject list thingie.

Also adds a rudimentary test for this in IntegrationTests.

I think this works? :^)

* Fixes filter(type="color",...) error, other weird color bugs

This PR also includes a parity note: I am not sure how BYOND handles setting a color filter when an atom already has a color var set to a color string (a ColorMatrix crashes in BYOND, however). We can investigate later if it's important :^)

EDIT: Removed hack once I figured out what the heck I was missing in my serialization class thing for ColorMatrix

* Refactors /atom.color to hold its filter separately from others

This is the... simplest thing I could think of for handling the fact that an atom can get a color matrix in two ways: one from a filter, and one from assigning to /atom.color.

Also includes some patches done in response to reviews on the PR by people who aren't Wixoa. Thanks everybody.

EDIT: Removes random debug print I left in there :^)

* Fixes improper sRGB <-> LRGB mapping for ColorMatrix

Eventually we're going to want a better solution to this, but this should be fine for now.

Also edited the values for the color matrix demo so that it will actually display any discrepancies caused by colour-space-related hooplah.

* Does the Wixoa review for color matrices

Is it over? Am I free?