Improves error emission in switch blocks, adds OD3201
Turns out, BYOND actually is OK with "else if" appearing in a switch block, but takes it to mean an 'else' of the switch, followed by an if-elseif-else ladder within that case block.

Beestation apparently uses this in production, but it's a very odd and ambiguous style (Ike told me he had to fix a bug caused by this recently in Para) so I've made a style warning for it.

This patch also includes some auditing of DMParser Error() emissions so that switch case parsing throws exceptions less often (which reduces code skipping).

I do not love how git submodules work

edit: was missing one comma