Optimize the passing of arguments between proc calls (#1203)
* Refactor DreamProcArguments
Ordered arguments are now held as a pooled array

* Refactor DreamProcArguments part 2
DreamProcArguments is now a ref struct that holds a ReadOnlySpan pointing to an array of DreamValues (a DMProc's stack in a majority of cases).

To allow for this they are no longer pushed to the stack and are instead operands to opcodes that call procs. The interpreter will also map named arguments to their position instead of putting them in a dictionary.

* Add color parameter to animate()

* Found out what's going on there
Ordered arguments were being overwritten to null by named arguments

* Revert DreamValue changes

* Use Array.Fill()

* Fix gradient()
Made it its own opcode. It has its own special argument handling.

* gradient() can take an arglist()

* Fix Gradient opcode disassembly

* Update OpenDreamRuntime/DreamConnection.cs

* Update OpenDreamRuntime/Procs/DMProc.cs