Fix ACZ root path being wrong if using relative JSON path
The JSON path is relative to the original working directory of the file. However the server switches working directory during initialization of DreamResourceManager, which invalidates the old path. This old path was still passed to ACZ, so using a relative JSON path would make ACZ unable to find rsc files.

The path is now turned absolute before the working directory switch occurs.