HtmlParser: Support named entities in body text
This gives HtmlParser the ability to support named entities in body text.
This is primarily useful to clean up text such as "&quot;", "&apos;" ,and "&#33;".

This was tested with [Tomeno's OpenDream branch of Console](https://github.com/Tomeno/console/tree/opendream_cope) but with changes to interface/default.dmf reverted and then the computer_docked window deleted. When operating a computer and "> "-commands are typed, the window no longer shows HTML entities with this patch.

...I'm sure there are simpler ways to test this, but that's where I found the issue.