Implement Byond_CRASH()
This relies on setjmp()/longjmp(), so I had to add C++ code to use those functions. This means I had to add cc to link in the C++ code. Send help.

We now run all call_exts() through the trampoline library so it can setjmp() before jumping to the target module.